+++ src/DotRecast.Recast.Toolset/Tools/RcCrowdTool.cs
@@ -297,7 +297,7 @@ namespace DotRecast.Recast.Toolset.Tools
             RcVec3f vel = RcVec3f.Subtract(tgt, pos);
             vel.Y = 0.0f;
             vel = RcVec3f.Normalize(vel);
-            return vel.Scale(speed);
+            return vel * speed;
         }
 
         public long GetCrowdUpdateTime()