+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Removed
 - Removed RcVecUtils.Dot()
+- Removed RcVecUtils.Scale()
 
 ### Special Thanks
 - [@Doprez](https://github.com/Doprez)